+++ CHANGELOG.md
@@ -12,7 +12,10 @@
 - Bug #155: Fix `Sort` configuration preparation (@vjik)
 - Bug #155: Fix same named order fields in `Sort` were not overriding previous ones (@vjik)
 - New #158: Add methods `PaginatorInterface::isSortable()` and `PaginatorInterface::withSort()` (@vjik)
-- New #158: Add `Sort::hasFieldInConfig()` method that check for the presence of a field in the sort config (@vjik)
+- Chg #159: Replace `withNextPageToken()` and `withPreviousPageToken()` of `PaginatorInterface` with `withToken()`,
+  `getNextPageToken()` with `getNextToken()`, `getPreviousPageToken()` with `getPreviousToken()`, and add `getToken()`.
+  These methods use new `PageToken` class (@vjik)
+- New #160: Add `Sort::hasFieldInConfig()` method that check for the presence of a field in the sort config (@vjik)
 
 ## 1.0.1 January 25, 2023